docs: add Lorem component
