+++ packages/docs/components/CodeBlock.jsx
@@ -9,23 +9,22 @@ import * as ReactBeautifulDND from 'react-beautiful-dnd';
 import * as ReactDND from 'react-dnd';
 import * as ReactDNDHtml5backend from 'react-dnd-html5-backend';
 import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live';
-import { loremIpsum } from 'lorem-ipsum';
 import * as ReactMovable from 'react-movable';
 import * as ReactTable from 'react-table';
 import { AutoSizer } from 'react-virtualized';
 import { codeBlockLight, codeBlockDark } from '../prism-themes/tonic-ui';
 import FontAwesomeIcon from './FontAwesomeIcon';
 import EditableTag from './EditableTag';
+import Lorem from './Lorem';
 
 const thirdPartyComponents = {
   AutoSizer,
-  FontAwesomeIcon,
   ReactBeautifulDND,
   ReactDND,
   ReactDNDHtml5backend,
   ReactMovable,
-  update,
-  ...ReactTable,
+  update, // XXX: rename to immutableUpdate
+  ...ReactTable, // XXX: rename to ReactTable.xxx
 };
 
 const {
@@ -192,36 +191,9 @@ const CodeBlock = ({
     scope: {
       ...styledUIComponents,
       ...thirdPartyComponents,
+      FontAwesomeIcon,
       EditableTag,
-      Lorem: ({
-        count = 1,
-        ...rest
-      }) => {
-        const html = loremIpsum({
-          count,
-          units: 'paragraphs',
-          sentenceLowerBound: 5,
-          sentenceUpperBound: 15,
-          paragraphLowerBound: 3,
-          paragraphUpperBound: 7,
-          format: 'html',
-        });
-
-        return (
-          <Box
-            dangerouslySetInnerHTML={{ __html: html }}
-            css={{
-              '> *:first-child': {
-                marginTop: 0,
-              },
-              '> *:last-child': {
-                marginBottom: 0,
-              },
-            }}
-            {...rest}
-          />
-        );
-      },
+      Lorem,
       css,
       mdx,
       tmicons,
